pyproject.toml
[tool.poetry]
name = "diplom"
version = "0.1.0"
description = ""
authors = ["Rezvanceva <elle81998@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
django = "^4.1.5"
envparse = "^0.2.0"
path = "^16.6.0"
psycopg2-binary = "^2.9.5"
gunicorn = "^20.1.0"
ansible-vault-win = "^1.0.0"
serializers = "^0.2.4"
djangorestframework = "^3.14.0"
cfgv = "^3.3.1"
generics = "^6.0.0"
response = "^0.5.0"
django-health-check = "^3.17.0"
views = "^0.3"
social-auth-app-django = "^5.0.0"
django-filter = "^22.1"
django-extensions = "^3.2.1"
pydantic = "^1.10.4"
basemodel = "^20190604.1625"
register = "^0.1"
pytest = "^7.2.1"
freezegun = "^1.2.2"
pytest-factoryboy = "^2.5.1"
mypy = "^1.0.0"
apiclient = "^1.0.4"
objects = "^0.3.1"
serializer = "^0.2.1"


[tool.poetry.group.dev.dependencies]
ansible-vault-win = "^1.0.0"
django-extensions = "^3.2.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
